Crate relay

Relay is a standalone service for scrubbing personal information and improving event submission response time. It acts as a middle layer between your application and Sentry.io. For information on setting up Relay in front of Sentry, see:

Relay documentation

Event Ingestion

Relay is also the primary event ingestion system at Sentry. When compiled with the processing feature and configured for ingestion, Relay ships functionality for:

Acknowledging and handling web requests on store endpoints

Applying rate limits and inbound data filters

Data scrubbing of event payloads and attachments

Validation and normalization of malformed or outdated data

Forwarding of all information into the event processing pipeline

Feature Flags

ssl (default): Enables SSL support using native-tls.

processing: Includes event ingestion and processing functionality. This should only be specified when compiling Relay as Sentry service. Standalone Relays do not need this feature.

Workspace Crates

Relay is split into the following workspace crates:

relay: Main entry point and command line interface.

relay-auth: Authentication and crypto.

relay-aws-extension: AWS extension implementation for Sentry’s AWS Lambda layer.

relay-cabi: C-bindings for exposing functionality to Python.

relay-common: Common utilities and crate re-exports.

relay-config: Configuration for the CLI and server.

relay-crash: Crash reporting for the Relay server.

relay-ffi: Utilities for error handling in FFI bindings.

relay-ffi-macros: Macros for error handling in FFI bindings.

relay-filter: Inbound data filters.

relay-general: Event protocol, normalization and data scrubbing.

relay-log: Error reporting and logging.

relay-metrics: Metrics protocol and processing.

relay-quotas: Sentry quotas and rate limiting.

relay-redis: Pooled Redis and Redis cluster abstraction.

relay-sampling: Dynamic sampling functionality.

relay-statsd: High-level StatsD metric client for internal measurements.

relay-server: Endpoints and services.

relay-system: Foundational system components for Relay’s services.

relay-test: Helpers for testing the web server and services.

Tools

Complementary to the Relay binary, the workspace contains a set of tools to interface with some of Relay’s functionality or to export information:

document-metrics: Generate documentation for metrics.

generate-schema: Generate an event payload schema file.

process-event: Process a Sentry event payload.

scrub-minidump: Scrub PII from a Minidump file.

Modules

cli 🔒

cliapp 🔒

setup 🔒

utils 🔒

Functions

main