Crate relay

Structs

BounceStatus

Represents a bounce status object nested within the profile.

RelayAddress

Represents a Relay email address object returned by the Relay API.

RelayClient

Represents a client for the Relay API.

RelayProfile

Represents a Relay user profile returned by the Relay API.

Enums

Error

RelayApiError

Type Aliases

ApiResult

Result